Refactored the code in line with new WLF best practice
