+++ webviz_subsurface/plugins/_group_tree/_views/__init__.py
@@ -1,0 +1,7 @@
+from ._group_tree_view import (
+    GroupTreeView,
+    GroupTreeViewElement,
+    ViewControls,
+    ViewFilters,
+    ViewOptions,
+)
