+++ webviz_subsurface/plugins/_group_tree/_views/_group_tree_view/_view_element.py
